requirements.txt
djangorestframework
django-cors-headers
drf-yasg
pillow
djoser
pytils
django-ckeditor